# docs/swagger.json
{
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "This is a simple API to manage goods for the Chaika app.",
    "title": "ChaikaGoods API",
    "contact": {
      "name": "Chaika",
      "email": "chaika.contact@yandex.ru"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "0.0.2"
  },
  "host": "127.0.0.1:8080",
  "basePath": "/",
  "paths": {
    "/api/v1/Templates": {
      "post": {
        "description": "Add a new Template of products to the database",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Templates"
        ],
        "summary": "Add Template",
        "parameters": [
          {
            "description": "Template details",
            "name": "Template",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/github_com_Chaika-Team_ChaikaGoods_internal_handler_schemas.AddTemplateRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/github_com_Chaika-Team_ChaikaGoods_internal_handler_schemas.AddTemplateResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/github_com_Chaika-Team_ChaikaGoods_internal_handler_schemas.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/github_com_Chaika-Team_ChaikaGoods_internal_handler_schemas.ErrorResponse"
            }
          }
        }
      }
    },
    "/api/v1/Templates/search": {
      "get": {
        "description": "Search for Templates",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Templates"
        ],
        "summary": "Search Template",
        "parameters": [
          {
            "type": "string",
            "description": "Search query",
            "name": "query",
            "in": "query"
          },
          {
            "type": "integer",
            "description": "Limit",
            "name": "limit",
            "in": "query",
            "required": true
          },
          {
            "type": "integer",
            "description": "Offset",
            "name": "offset",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/github_com_Chaika-Team_ChaikaGoods_internal_handler_schemas.SearchTemplatesResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/github_com_Chaika-Team_ChaikaGoods_internal_handler_schemas.ErrorResponse"
            }
          }
        }
      }
    },
    "/api/v1/Templates/{id}": {
      "get": {
        "description": "Get Template details by its ID",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Templates"
        ],
        "summary": "Get Template by ID",
        "parameters": [
          {
            "type": "integer",
            "description": "Template ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/github_com_Chaika-Team_ChaikaGoods_internal_handler_schemas.GetTemplateByIDResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/github_com_Chaika-Team_ChaikaGoods_internal_handler_schemas.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/github_com_Chaika-Team_ChaikaGoods_internal_handler_schemas.ErrorResponse"
            }
          }
        }
      }
    },
    "/api/v1/products": {
      "get": {
        "description": "Get all products from the database",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "products"
        ],
        "summary": "Get all products",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/github_com_Chaika-Team_ChaikaGoods_internal_handler_schemas.GetAllProductsResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/github_com_Chaika-Team_ChaikaGoods_internal_handler_schemas.ErrorResponse"
            }
          }
        }
      },
      "put": {
        "description": "Update product details in the database",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "products"
        ],
        "summary": "Update product",
        "parameters": [
          {
            "description": "Product details",
            "name": "product",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/github_com_Chaika-Team_ChaikaGoods_internal_handler_schemas.UpdateProductRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/github_com_Chaika-Team_ChaikaGoods_internal_handler_schemas.UpdateProductResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/github_com_Chaika-Team_ChaikaGoods_internal_handler_schemas.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/github_com_Chaika-Team_ChaikaGoods_internal_handler_schemas.ErrorResponse"
            }
          }
        }
      },
      "post": {
        "description": "Add a new product to the database",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "products"
        ],
        "summary": "Add product",
        "parameters": [
          {
            "description": "Product details",
            "name": "product",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/github_com_Chaika-Team_ChaikaGoods_internal_handler_schemas.CreateProductRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/github_com_Chaika-Team_ChaikaGoods_internal_handler_schemas.CreateProductResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/github_com_Chaika-Team_ChaikaGoods_internal_handler_schemas.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/github_com_Chaika-Team_ChaikaGoods_internal_handler_schemas.ErrorResponse"
            }
          }
        }
      }
    },
    "/api/v1/products/{id}": {
      "get": {
        "description": "Get product details by its ID",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "products"
        ],
        "summary": "Get product by ID",
        "parameters": [
          {
            "type": "integer",
            "description": "Product ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/github_com_Chaika-Team_ChaikaGoods_internal_handler_schemas.GetProductByIDResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/github_com_Chaika-Team_ChaikaGoods_internal_handler_schemas.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/github_com_Chaika-Team_ChaikaGoods_internal_handler_schemas.ErrorResponse"
            }
          }
        }
      },
      "delete": {
        "description": "Delete a product from the database",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "products"
        ],
        "summary": "Delete product",
        "parameters": [
          {
            "type": "integer",
            "description": "Product ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/github_com_Chaika-Team_ChaikaGoods_internal_handler_schemas.DeleteProductResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/github_com_Chaika-Team_ChaikaGoods_internal_handler_schemas.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/github_com_Chaika-Team_ChaikaGoods_internal_handler_schemas.ErrorResponse"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "github_com_Chaika-Team_ChaikaGoods_internal_handler_schemas.AddTemplateRequest": {
      "description": "Запрос на добавление пакета",
      "type": "object",
      "properties": {
        "Template": {
          "description": "Сведения о новом пакете",
          "allOf": [
            {
              "$ref": "#/definitions/github_com_Chaika-Team_ChaikaGoods_internal_handler_schemas.TemplateSchema"
            }
          ]
        }
      }
    },
    "github_com_Chaika-Team_ChaikaGoods_internal_handler_schemas.AddTemplateResponse": {
      "description": "Ответ на запрос на добавление пакета",
      "type": "object",
      "properties": {
        "id": {
          "description": "ID созданного пакета",
          "type": "integer"
        }
      }
    },
    "github_com_Chaika-Team_ChaikaGoods_internal_handler_schemas.CreateProductRequest": {
      "description": "Запрос на добавление продукта",
      "type": "object",
      "properties": {
        "product": {
          "$ref": "#/definitions/github_com_Chaika-Team_ChaikaGoods_internal_handler_schemas.ProductSchema"
        }
      }
    },
    "github_com_Chaika-Team_ChaikaGoods_internal_handler_schemas.CreateProductResponse": {
      "description": "Ответ на запрос на добавление продукта",
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        }
      }
    },
    "github_com_Chaika-Team_ChaikaGoods_internal_handler_schemas.DeleteProductResponse": {
      "description": "Ответ на запрос на удаление продукта",
      "type": "object"
    },
    "github_com_Chaika-Team_ChaikaGoods_internal_handler_schemas.ErrorResponse": {
      "type": "object",
      "properties": {
        "code": {
          "description": "Код ошибки",
          "type": "integer"
        },
        "message": {
          "description": "Сообщение об ошибке",
          "type": "string"
        }
      }
    },
    "github_com_Chaika-Team_ChaikaGoods_internal_handler_schemas.GetAllProductsResponse": {
      "description": "Ответ на запрос на получение всех продуктов",
      "type": "object",
      "properties": {
        "products": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/github_com_Chaika-Team_ChaikaGoods_internal_handler_schemas.ProductSchema"
          }
        }
      }
    },
    "github_com_Chaika-Team_ChaikaGoods_internal_handler_schemas.GetProductByIDResponse": {
      "description": "Ответ на запрос на получение продукта по его ID",
      "type": "object",
      "properties": {
        "product": {
          "$ref": "#/definitions/github_com_Chaika-Team_ChaikaGoods_internal_handler_schemas.ProductSchema"
        }
      }
    },
    "github_com_Chaika-Team_ChaikaGoods_internal_handler_schemas.GetTemplateByIDResponse": {
      "description": "Ответ на запрос на получение пакета по его ID",
      "type": "object",
      "properties": {
        "Template": {
          "$ref": "#/definitions/github_com_Chaika-Team_ChaikaGoods_internal_handler_schemas.TemplateSchema"
        }
      }
    },
    "github_com_Chaika-Team_ChaikaGoods_internal_handler_schemas.ProductSchema": {
      "type": "object",
      "properties": {
        "description": {
          "type": "string"
        },
        "id": {
          "type": "integer"
        },
        "imageurl": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "price": {
          "type": "number"
        }
      }
    },
    "github_com_Chaika-Team_ChaikaGoods_internal_handler_schemas.SearchTemplatesResponse": {
      "description": "Ответ на запрос на поиск пакетов",
      "type": "object",
      "properties": {
        "Templates": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/github_com_Chaika-Team_ChaikaGoods_internal_handler_schemas.TemplateSchema"
          }
        }
      }
    },
    "github_com_Chaika-Team_ChaikaGoods_internal_handler_schemas.TemplateContentSchema": {
      "type": "object",
      "properties": {
        "product_id": {
          "type": "integer"
        },
        "quantity": {
          "type": "integer"
        }
      }
    },
    "github_com_Chaika-Team_ChaikaGoods_internal_handler_schemas.TemplateSchema": {
      "type": "object",
      "properties": {
        "content": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/github_com_Chaika-Team_ChaikaGoods_internal_handler_schemas.TemplateContentSchema"
          }
        },
        "description": {
          "type": "string"
        },
        "id": {
          "type": "integer"
        },
        "template_name": {
          "type": "string"
        }
      }
    },
    "github_com_Chaika-Team_ChaikaGoods_internal_handler_schemas.UpdateProductRequest": {
      "description": "Запрос на обновление продукта",
      "type": "object",
      "properties": {
        "product": {
          "$ref": "#/definitions/github_com_Chaika-Team_ChaikaGoods_internal_handler_schemas.ProductSchema"
        }
      }
    },
    "github_com_Chaika-Team_ChaikaGoods_internal_handler_schemas.UpdateProductResponse": {
      "description": "Ответ на запрос на обновление продукта",
      "type": "object"
    }
  }
}